norme ok

+++ Philo_three/include/philo.h
@@ -22,7 +22,7 @@
 # include <semaphore.h>
 # include <signal.h>
 
-typedef struct	s_data
+typedef struct s_data
 {
 	int			nbr_of_philo;
 	int			time_to_die;
@@ -88,7 +88,7 @@ int				free_all(t_data *values);
 *** Libft
 */
 
-int				ft_atoi(const char *str);
+long long		ft_atoi(const char *str);
 char			*ft_itoa(int n);
 char			*ft_strdup(const char *s1);
 size_t			ft_strlen(const char *s);
